+++ test/DataTable/DataTable.test.tsx
@@ -1,3 +1,8 @@
+import { TextDecoder, TextEncoder } from "util";
+
+global.TextEncoder = TextEncoder as unknown as typeof global.TextEncoder;
+global.TextDecoder = TextDecoder as unknown as typeof global.TextDecoder;
+
 import { render, fireEvent } from "@testing-library/react";
 import { renderHook } from "@testing-library/react-hooks/server";
 import "@testing-library/jest-dom";
@@ -5,7 +10,6 @@ import React from "react";
 import { ReactDataTable, createReactDataTableColumnHelper, useReactDataTable } from "src";
 
 global.React = React; // this also works for other globally available libraries
-
 function padTo2Digits(num: number) {
   return num.toString().padStart(2, "0");
 }